feat(onboarding): migrate display-name screen into onboarding flow for #356
Rename PickUsernameScreen into the onboarding DisplayNameScreen as step 2,
implement the setDisplayName action (wraps setUsername + records
onboarding.last_visited_path), and drop the legacy HomeScreen username
gate plus the now-orphaned PICK_USERNAME route/screen.
