+++ src/libs/Navigation/linkingConfig/config.ts
@@ -18,8 +18,6 @@ const config: LinkingOptions<RootStackParamList>['config'] = {
     [SCREENS.FORGOT_PASSWORD]: ROUTES.FORGOT_PASSWORD,
     [SCREENS.FORCE_UPDATE]: ROUTES.FORCE_UPDATE,
 
-    [SCREENS.PICK_USERNAME]: ROUTES.PICK_USERNAME,
-
     // Sidebar
     [NAVIGATORS.BOTTOM_TAB_NAVIGATOR]: {
       path: ROUTES.ROOT,
